requirements.txt
pandas==1.5.0
numpy==1.23.4
SQLAlchemy==1.4.41
psycopg2-binary==2.9.4
python-dotenv==0.21.0
loguru==0.6.0